move dot-template-props
