+++ src/app/portlets/dot-templates/dot-template-designer/dot-template-props/dot-template-props.component.ts
